app/backend/requirements.txt
azure-identity==1.16.0
azure-search-documents==11.4.0
quart==0.19.6
openai[datalib]==1.30.5
uvicorn[standard]==0.30.1
aiohttp==3.9.5
numpy==1.26.4
pyyaml==6.0.1
postgres==4.0
beautifulsoup4==4.12.3
requests==2.32.3
redis==5.0.4